+++ cynic-introspection/src/lib.rs
@@ -27,13 +27,33 @@
 //! assert_eq!(schema.query_type, "Root");
 //! ```
 //!
+//! ### GraphQL Versions
+//!
+//! GraphQL servers currently commonly support two different versions of the GraphQL
+//! specification:
+//!
+//! - [The June 2018 Specification][3]
+//! - [The October 2021 Specification][4]
+//!
+//! The fields available for introspection differ between these two versions, so cynic
+//! provides two different introspection queries in the modules [query2018] and [query2021]
+//! respectively.
+//!
+//! It also provides [CapabilityDetectionQuery], a query which can determine which version
+//! of the specification a GraphQL server supports.
+//!
 //! [1]: http://spec.graphql.org/October2021/#sec-Introspection
 //! [2]: https://cynic-rs.dev
+//! [3]: http://spec.graphql.org/June2018
+//! [4]: http://spec.graphql.org/October2021
 
+mod detection;
 pub mod query2018;
 pub mod query2021;
 mod schema;
 
+#[doc(inline)]
+pub use detection::{CapabilityDetectionQuery, SpecificationVersion};
 pub use query2018 as query;
 #[doc(inline)]
 pub use query2018::{DirectiveLocation, IntrospectionQuery};